Subindo telaLista completa

+++ Cachorros/app/src/main/java/com/example/cachorros/Cachorro.kt
@@ -1,3 +1,3 @@
 package com.example.cachorros
 
-data class Cachorro (val id:Int, val raca:String, val precoMedio:Int, val indicadoCriancas:Boolean)
+data class Cachorro (val id:Int, val raca:String, val precoMedio:Double, val indicadoCriancas:Boolean)
